melon/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
Django==4.1.7
easyocr==1.6.2
exceptiongroup==1.1.1
filelock==3.10.3
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.2
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hydra-core==1.3.2
idna==3.4
imageio==2.26.1
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.10
jax==0.4.6
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
labelme2coco==0.2.4
lazy_loader==0.2
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.1
pause==0.3
Pillow==9.4.0
pixellib==0.7.1
platformdirs==3.1.1
playsound==1.3.0
portalocker==2.7.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pyclipper==1.3.0.post4
pycparser==2.21
pydot==1.4.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.12
scikit-image==0.20.0
scipy==1.10.1
selenium==3.14.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminaltables==3.1.10
tifffile==2023.3.21
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
xmljson==0.2.1
yacs==0.1.8
